requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
argon2-cffi==21.3.0
    # via notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.5.2
    # via django
asttokens==2.0.8
    # via stack-data
atomicwrites==1.4.0
    # via pytest
attrs==21.4.0
    # via
    #   jsonschema
    #   pytest
autopep8==1.6.0
    # via -r requirements.in
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via nbconvert
bleach==5.0.1
    # via nbconvert
certifi==2022.5.18.1
    # via requests
cffi==1.15.1
    # via argon2-cffi-bindings
charset-normalizer==2.0.12
    # via requests
colorama==0.4.4
    # via
    #   ipython
    #   pytest
debugpy==1.6.3
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
django==4.0.4
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.13.1
    # via -r requirements.in
entrypoints==0.4
    # via jupyter-client
executing==1.1.1
    # via stack-data
faker==13.13.0
    # via -r requirements.in
fastjsonschema==2.16.2
    # via nbformat
idna==3.3
    # via requests
importlib-metadata==5.0.0
    # via nbconvert
iniconfig==1.1.1
    # via pytest
ipykernel==6.16.0
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   notebook
    #   qtconsole
ipython==8.5.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipython-genutils==0.2.0
    # via
    #   notebook
    #   qtconsole
ipywidgets==8.0.2
    # via jupyter
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via
    #   nbconvert
    #   notebook
jsonpath==0.82
    # via -r requirements.in
jsonschema==4.16.0
    # via nbformat
jupyter==1.0.0
    # via -r requirements.in
jupyter-client==7.4.2
    # via
    #   ipykernel
    #   jupyter-console
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.4.4
    # via jupyter
jupyter-core==4.11.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-widgets==3.0.3
    # via ipywidgets
markupsafe==2.1.1
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mistune==2.0.4
    # via nbconvert
mpmath==1.2.1
    # via sympy
nbclient==0.7.0
    # via nbconvert
nbconvert==7.2.1
    # via
    #   jupyter
    #   notebook
nbformat==5.7.0
    # via
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.6
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   notebook
notebook==6.4.12
    # via jupyter
numpy==1.22.4
    # via pandas
packaging==21.3
    # via
    #   ipykernel
    #   nbconvert
    #   pytest
    #   qtpy
pandas==1.4.2
    # via -r requirements.in
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pickleshare==0.7.5
    # via ipython
pluggy==1.0.0
    # via pytest
prometheus-client==0.14.1
    # via notebook
prompt-toolkit==3.0.31
    # via
    #   ipython
    #   jupyter-console
psutil==5.9.2
    # via ipykernel
psycopg2-binary==2.9.3
    # via -r requirements.in
pure-eval==0.2.2
    # via stack-data
py==1.11.0
    # via pytest
pycodestyle==2.8.0
    # via autopep8
pycparser==2.21
    # via cffi
pycryptodome==3.17
    # via -r requirements.in
pygments==2.13.0
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
    #   qtconsole
pyjwt==2.5.0
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
pytest==7.1.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   faker
    #   jupyter-client
    #   pandas
python-dotenv==0.20.0
    # via -r requirements.in
pytz==2022.1
    # via
    #   djangorestframework
    #   pandas
pywin32==304
    # via jupyter-core
pywinpty==2.0.8
    # via terminado
pyzmq==24.0.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   qtconsole
qtconsole==5.3.2
    # via jupyter
qtpy==2.2.1
    # via qtconsole
requests==2.27.1
    # via -r requirements.in
send2trash==1.8.0
    # via notebook
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
stack-data==0.5.1
    # via ipython
sympy==1.10.1
    # via -r requirements.in
terminado==0.16.0
    # via notebook
tinycss2==1.1.1
    # via nbconvert
toml==0.10.2
    # via autopep8
tomli==2.0.1
    # via pytest
tornado==6.2
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
traitlets==5.4.0
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
tzdata==2022.1
    # via django
urllib3==1.26.9
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
widgetsnbextension==4.0.3
    # via ipywidgets
zipp==3.9.0
    # via importlib-metadata
